Move Task model to workflows submodule
